pyproject.toml
[tool.poetry]
name = "test-service-for-product-management"
version = "0.1.0"
description = ""
authors = ["RomanL91 <dev_roman91@mail.ru>"]
readme = "README.md"
packages = [{include = "test_service_for_product_management"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.1"
black = "^24.4.2"
django-mptt = "^0.16.0"
django-mptt-admin = "^2.6.2"
pillow = "^10.3.0"
django-cors-headers = "^4.3.1"
uvicorn = "^0.30.0"
django-flat-json-widget = "^0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
